chore(app): Rename apm.ts to sentry.ts
