+++ packages/openneuro-app/src/scripts/sentry.ts
